Ticket #5755: Move common abilities to a module to avoid duplication
